Add endpoint to get like count for a post
Implemented a new HTTP GET endpoint in the `LikesController` to retrieve the like count for a specific post by its ID. The `GetLikeCount` method checks for the existence of the post and returns a 404 Not Found response if it does not exist. If the post is found, it counts and returns the number of likes associated with that post.